[new-env] Lex scope of switch should not apply to the switch discriminant
Summary:
Changelog: [internal]

The inside of the switch forms its own lexical scope, which the discriminant should not be part of. Inside the switch cases, we might read a variable that has the name as the variable in discriminant. This is OK because the existing logic already handles the shadowing correctly. However, the [last read](https://github.com/facebook/flow/blob/6e86f8d8190d1c9632886628559630a9e005bf58/src/analysis/env_builder/name_resolver.ml#L2074-L2077) of the discriminant inside the switch cases still uses the environment inside the switch cases, which, in the case of shadowing, caused the read entry of the shadowed variable to point to the def inside the switch.

This diff performs a reset after visiting the switch cases using the initial env for the discriminant, so that outside of the switch, the variables in the discriminant is pointing the correct def again.

Reviewed By: jbrown215

Differential Revision: D34805836

fbshipit-source-id: 185f5b004eb2d6743e28e941bac092d31fdd8c9a