[new-env] Correctly model lexical bindings within switch
Summary:
Changelog: [internal]

Lexical bindings within switch are very awkward to use and model. When we have `let x = ...` in one switch case, the name `x` is not usable at all in all other switch cases. In all other cases, it should behave like an undeclared-or-skipped variable: definitely not declared in cases before, and the declaration might be skipped in the cases after.

As far as behaviors correctly fixed in this diff are concerned, this means the following errors :

1. `let x` in two different cases should be an rebinding error
2. Any other types of references of `x` is `EReferencedBeforeDeclaration` in all other cases than the case that introduces the `x` lexical binding.

Before this diff, the new env is not correctly modeling the behavior. Lucky cases like

```lang=flow
switch (...) {
  case 1:
    let x = 0;
  case 2:
    let x = 0;
}
```

just happen to error the right thing, because we merge the env when there is fall through, and the merge produces `phi[undeclared, loc]`, which is treated as `not is_undeclared`. In cases where there is a break like

```lang=flow
switch (...) {
  case 1:
    let x = 0;
    break;
  case 2:
    let x = 0;
}
```

It will not error at all, since the fallthrough env is not merged into the second case env, and the statements in the second case think that declaration of `x` never happens before, and it should be the source of the declaration.

Unfortunately, none of the current `write_state` can correctly model this behavior. To correctly error on `EReferencedBeforeDeclaration`, we need to mark such variables as undeclared, but we also use the state of undeclared to detect rebinding errors, which will cause type 1 error to disappear.

To fix this issue, we need to introduce a new kind, which I call `DeclaredButSkipped` (inspired by the error message of `EReferencedBeforeDeclaration`: "because the declaration either comes later or was skipped"). For each case, we reset all vals of previous cases' lexical binding to `DeclaredButSkipped` to correctly model the behavior. `Val.is_undeclared_or_skipped` is added in addition to `Val.is_undeclared` so that we can call `Val.is_undeclared_or_skipped` for the purpose of finding forward reference and call `Val.is_undeclared` to find rebinding errors.

Reviewed By: jbrown215

Differential Revision: D34907831

fbshipit-source-id: a911846b1e60224cb7b62d01361603d02e0a26d0