[scope] Fix scope of default expressions in function param
Summary:
Changelog: [fix] Correctly determine the scope of default expressions in function param to be outside of the function scope.

We are incorrectly determining the scope of default expressions in the scope builder. In the following example

https://github.com/facebook/flow/blob/b3f83f1e127be72a455f6e18459d683d42b0f4e3/tests/binding/scope.js#L98-L104

The existing incorrect behavior determines that `a` in `() => a` reads the value of `a=0` inside the function rather than the correct `a=""` outside of the function. This caused us to miss a type error in the new environment.

This diff corrects the behavior by splitting the visit of default expressions and function identifiers: one outside of function scope and one inside.

Reviewed By: jbrown215

Differential Revision: D34950010

fbshipit-source-id: 8c5187ab27fa704cd23f1aeb6f364532898b6b51