[new-env] Error on writes to unknown globals
Summary:
Changelog: [internal]

In the old env, we check for writes to undefined global by first trying to find the entry in the scope, if we couldn't find it, we call `Flow_js.get_builtin` to resolve it.

Right now in new env, we don't have the same check. In addition, we even have no reliable way to tell whether a name is a global. Information passed from scope builder, ssa builder, and name resolver all record names read only.

This diff fixes the problem by forwarding `unbound_names` computed during name_resolver building into `env_info`. Then when we are setting a variable, and only when the write is a provider (ensure we error on the global in the first write), we check whether the name is in `unbound_names`, and if it is, try to resolve by `Flow_js.get_builtin`.

Reviewed By: jbrown215

Differential Revision: D34948089

fbshipit-source-id: 22f85dc9f94d83ebcb7d6a31d6012650faecd8a6