Use more specific loc for the imported name in `import typeof *`
Summary:
Changelog: [internal]

For `import typeof * as Name from ...`, we currently use the loc of the entire import statement as the loc for the imported `Name`. This is problematic for the new env, since we attribute the write to the more specific loc that just contains `Name`. With this change, the imported typeof namespace can be correctly unified with the initial open tvar.

Reviewed By: gkz

Differential Revision: D35021348

fbshipit-source-id: 9195d79f5a69f348a708a23ea5ab2e89c3fd67ab