[new-env] Recover missing `ObjectSpread` op in new-env
Summary:
Changelog: [internal]

In the new env, illegal spread is already detected in `Flow.widen_obj_type` before the call to `Flow.flow` in the next line. (Because the spread value is likely to be `OpenT` due to unification with the open tvar given to each write during env init).

Therefore, it's better to give it the same use_op instead of the generic unknown use, so we can have the same detailed kind of error messages in the new env.

Reviewed By: jbrown215

Differential Revision: D35026034

fbshipit-source-id: 53bf5bf12b85b3082c59ac16e7e48cc850f99e72