more precisely model entrance state of catch block
Summary:
Havocing over the exit state of the try block caused variables that were uninitialized before the try block and initialized in the try block to become definitely initialized in the catch block. We now reset to the entrance state of the try block and havoc over that so that any variables that were uninitialized before the try remain uninitialized in the catch entrance state. For example:

```
function test(a, b) {
  var c;
  try {
    c = b();
  } catch {
    c; // c thought to be definitely initialized here
  };
  return c; // and thus here as well
}
```

After this fix we now know that `c` may be uninitialized in both the catch and the return.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D35054925

fbshipit-source-id: be5e8f4dc5b14c966d76c6da9bd2950d7c803634