Correctly model entrance state of finally block
Summary:
Like in the previous diff, havocing over the exit state of the catch block caused variables that were uninitialized before the catch block and initialized in the catch block to become definitely initialized in the finally block. We now reset to the entrance state of the try statement and havoc over that so that any variables that were uninitialized before the try/catch remain uninitialized in the finally entrance state. For example:
```
function test(a, b) {
  var c;
  try {
    c = b();
  } finally { c };
  return c;
```

After this fix we now know that `c` may be uninitialized in both the finally and the return.

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D35056958

fbshipit-source-id: 0d9cd2af6ddede4f7357a01558df92381f4cccec