fix adding suppressions for "cannot create element" errors
Summary:
"cannot create element" errors due to invalid children span the entire contents of the JSX children -- the character after the opening tag's `>`, through the character before the closing tag's `<`.

if the character after the `>` is a newline, this confuses tool's `findStartOfLine` function. that function walks backwards from the error location looking for a newline, so it thinks it found the start of the line but it's really stuck at the end of the line, causing it to insert the suppression within the children instead of before the children.

further, the "context" algorithm to decide whether we're within JSX and should insert `{/* $FlowFixMe ... */}` or in normal code and should insert `// $FlowFixMe ...` was wrong in this case. if the error is on the same line as part of the opening tag, then we need to insert the suppression before or within the opening tag. it currently incorrectly thinks it's within a JSX context.

Reviewed By: jbrown215

Differential Revision: D35079824

fbshipit-source-id: c3ab246ece54f58b439bd1d17f6fac381b5f6f59