Move all providers to shared memory, parallelize commit modules
Summary:
Flow's dependency graph models "depends on" relationships as arrows from files
to modules, and modules can have 0 or more provider files.

When files change, they might become new providers, or stop being a provider, or
both. The affected modules need to pick a new provider and the module's reverse
dependecies might need to be re-checked.

Before this diff, we tracked the list of all providers for a module in a hash table
in the server process's OCaml heap. Storing the providers here severely limited
parallelism. Furthermore, the all providers hash table consumed a lot of memory
for large repos.

This diff maintains these all providers lists in shared memory directly, as a
linked list of files. For example, a file module (M) points to a list of
provider files:

```
+---+---+---+
| M | * |...|
+---+---+---+
      |
  providers
      |     +--next--+     +--next--+
      v     |        v     |        v
      +---+---+---+  +---+---+---+  +---+---+---+
      | F | * |...|  | F | * |...|  | F | 0 |...|
      +---+---+---+  +---+---+---+  +---+---+---+
```

A haste module's (M) all providers list is more complicated. The tail of the
linked list is stored in the file's (F) parse entity's (E) latest data (P),
instead of in the file itself:

```
+---+---+---+   +---+---+---+                         +---+---+---+
| M | * |...|   | E | * |...|                         | E | * |...|
+---+---+---+   +---+---+---+                         +---+---+---+
      |         ^     |                               ^     |
  providers     |     +--+     M                      |     +--> etc.
      |       parse      |     ^   +--next--+       parse
      v         |        v     |   |        v         |
      +---+---+---+---+  +---+---+---+---+  +---+---+---+---+
      | F |   | * |...|  | P | * | * |...|  | F |   | * |...|
      +---+---+---+---+  +---+---+---+---+  +---+---+---+---+
```

(Note that the parse data P refers back to the haste module M as well.)

When files changes, we need to (a) update the all providers lists and (b)
re-pick providers for all dirty modules. Before, (a) was done through
calc_old_modules and calc_new_modules. With this diff, we do this work during
parsing itself, which saves a lot of time.

Because the all providers lists are in shared memory, we can re-pick providers
in parallel using workers, whereas before this work needed to be done in a
single-threaded manner. This also saves a lot of time.

Reviewed By: mroch

Differential Revision: D34990416

fbshipit-source-id: 68a0d8a78be1c250b6a352d863f47ec4eb3dcc39