Refactor logic for re-picking module providers
Summary:
Now that a module's "all providers" set is a list of file addresses, we can
avoid the get_file_addr_unsafe call here. In order to do this, we pass the set
of `File_key.t * file_addr` pairs to each module system's commit modules
implementation.

These implementations continue use the key to choose a provider --
distinguishing between mocks, impls, and decls -- but now return the file addr
instead of the key.

Reviewed By: jbrown215

Differential Revision: D34990675

fbshipit-source-id: c39730821f24f38007fb2b22249242be795e4699