Reference file module object from file object
Summary:
We can look up a file module through the shared hash table, but it's simpler to
store the address directly on the file. This is useful in later diffs to manage
a file module's providers. For example, when deleting a file, also removing the
file from the file module's providers.

Reviewed By: jbrown215

Differential Revision: D34932106

fbshipit-source-id: d55de0fd815b6895ce33529f39bf738dc5343eb2