Redeclared vars should be considered annotate if provider is annotated
Summary:
A `var` can be redeclar.ed multiple times, but the provider analysis will consider it annotated if one of the declarations is annotated. In that case, there is no reason to unify with the write slot, which is constrained from above by the providers.

Take this example:
```
var a: number = 3;
a;
var a = 'str';
```

When we `bind` the first declaration of `a` to the environment we set the type of the write slot to `number`, as provided by the annotation. Since that first declaration is also a provider, we do not need to flow `number` to the types of the provider, but we do check that `3` is compatible with the `number` annotation.

On the second declaration, which is not a provider and has no annotation, we will unify a synthetic "annot" that statement.ml creates that is an empty tvar with the write slot. Since that location is not a provider, we subtype that OpenT against the providers, which is just `number` from the annotation.

When we see the assignment of `str` to `a`, we would unify `str` with the synthetic tvar with a `number` upper bound, hitting a weird case in unification that would cause the error message to be flipped. Fixing that error message use_op has proven to be a massive pain, since many erro rmessages in our tests depend on very specific compositions of use ops.

This diff avoids that unification altogether by detecting that the binding is annotated somewhere, so instead of unifying with the synthetic annotation we just flow to it, and we get a normal error message back.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D35130782

fbshipit-source-id: 63fe16cfa2759bd517d9b0dc2d44e9509378e608