Add missing module generator functionality
Summary:
We received a user request to be able to make a custom error for missing modules when that missing module could be generated (e.g. by Relay). This would help the end user resolve the missing module error.

Rather than hardcoding support for some specific generators, I believe it is better to make this configurable instead. Even projects using the same generators (e.g. Relay) may use different commands to generate the relevant files. Additionally, there may be other generators that could also be supported.

Since all we have to go off of is the module name, we can't cover all generated modules, but for some which follow a certain pattern it is possible.

Here we just add a custom error message. In the future, if there is some way to do so (and there are no security concerns), it could be possible to also add a "quick fix" button to run the command.

Changelog: [feature] Added the `module.missing_module_generators` option, which can be used (multiple times) to specify `'regex' -> 'command'` pairs. When a module is missing, if its name matches one of the regexes, we will add a suggestion in the error message to run `command` to generate it (and resolve the error).

Reviewed By: SamChou19815

Differential Revision: D35132371

fbshipit-source-id: 6346920486e54c2539c99d82250904dbcff69e76