Reference file module object from file object
Summary:
We can look up a file module through the shared hash table, but it's simpler to
store the address directly on the file. This is useful in later diffs to manage
a file module's providers. For example, when deleting a file, also removing the
file from the file module's providers.

Reviewed By: jbrown215

Differential Revision: D34932106

fbshipit-source-id: 385322e40f426a11b17bc4d7cc84ff6f927eb5a5