Support tagged int field values in shared memory
Summary:
Currently, scanned heap objects must only contain addrs. If we want to store an
integer, we need to allocate a separate int64 object.

Because all objects are aligned, the low order bit of an addr will always be 0.
This diff teaches the marking pass and compaction to check the low bit of a
field before assuming it is an addr.

With this support in place, we can write non-address values into a field of a
scanned object, as long as the low bit is set.

Reviewed By: jbrown215

Differential Revision: D35208297

fbshipit-source-id: fc285b4d45d1fc21e69fe17d2db912cd39e40edb