[new-env] Loosely model this and super in new env to make sure their refinements can be tracked
Summary:
The name resolver and the new env API in the current form does not have the ability to accurately represent the definitions of `this` and `super`. The biggest challenge is their definitions never appear in the source text, so there is no obvious choice for the def loc. Choosing the function loc as loc would be possible if we only have `this`, but the presence of `super` and `arguments` (if we still want to support them) might choose the same location and complicates the effort.

This diff tries to partially solve the problem by modeling this and super as var bindings in the module scope. This is of course inaccurate, since this in functions and methods should shadow the this in the global scope, but it happens to serve our purpose well.

- `this` and `super` are keywords, and it's a syntax error to ever reassign them, so we don't need to worry about any illegal writes to them.
- All `this` and `super` appearing in the code will be reads, although a lot of them would be reading at the wrong def location. The wrong def location doesn't matter in type lookup, since it will be delegated to the old env, which looks up type by name.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D35134449

fbshipit-source-id: bb1380282cbb2f18ba1f94c7f536cec06a34554c