Ban `new` on functions
Summary:
Before this change, we allowed functions to act as constructors. This was done in order to support the pattern that existed before ES6 classes.

```
function C(a) {
  this.a = a;
}
const c = new C(1);
```

We are motivated to remove support for this pattern not only because it has been obviated by ES6 classes, but also because it relies on unsealed objects to model `this`, which we are are in the process of eliminating.

Changelog: [error] Banned usage of `new` on functions. Move usages of this pattern to use ES6 classes instead. If the pattern exists in third-party that cannot be changed directly, you can use a [declaration file](https://flow.org/en/docs/declarations/) to type the legacy pattern with a `declare class`.

Reviewed By: samwgoldman

Differential Revision: D35037868

fbshipit-source-id: 252e726212b0c09be4bbc77e505bc44082d18286