Rename phantom dependents -> phantom dependencies
Summary:
In Flow, we try to use these two terms carefully. If X has a "dependency" on Y, then X requires Y. In order to analyze X, we need to analyze Y first.

A "dependent" is the opposite. In the above, X is a "dependent" of Y. When Y changes, we might need to re-analyze X. This is the same thing as a "reverse dependency."

In the resolved requires record, we store "phantom dependencies" -- lists of file paths which do not exist, but if they did exist, would become our dependencies.

We might also say that a file having a phantom dependency on a "phantom file" F is a "phantom dependent" of F.

Reviewed By: mroch

Differential Revision: D35227735

fbshipit-source-id: 2db11f853574a215fb3c02dde8a129f6de398658