[new-env] Prevent merges of variables that have the same name but in different scopes
Summary:
We have an interesting test case that can cause a crash when we try to find the provider type of `foo` in `const {foo}`.

```
switch ('foo') {
    case 'foo': {
        const foo = 3;
        break;
    }
}

const {foo} = {foo:3};
```

Digging deeper, it turns out that it has nothing to do with provider analysis, but instead related in how we merge environments. In this example, the `break` causes an abrupt completion, which later causes its environment to be merged with the lexical environment inside switch. Since the case is wrapped with a block, **it causes a merge between `foo` in the program scope, and the `foo` in the scope of the case**. This produces an invalid `phi[undeclared, loc]` write state, which is assumed to be an existing declaration, and make `name_resolver` mark the outer foo declaration as duplicate and an non-assigning write. This caused a discrepancy between the provider analysis, which thinks the outer foo is its own provider, and the name resolver analysis, which thinks the outer foo is not an assigning write. The discrepancy caused a crash, since we failed to find the type of the write in loc_env. (Such problem is also possible in loops, which I included in test cases as well)

To fix this, we need to prevent merging together Vals from different scopes. I added `def_loc` to `Env.entry` to distinguish variables from different scopes, so they won't be accidentally merged together just because they happen to share the same name, which is expected to happen, since `merge_remote_env` can bring in a deeply nested env inside the current environment.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D35200086

fbshipit-source-id: f8f0eea6d364aef6a0f35ab98f1349907d01d3bd