fix version mismatch handling
Summary:
when a client connects to a server, they exchange a handshake to make sure they are the same binary. if they are not the same binary, we have to decide what to do: kill the server, or kill the client.

if the server is a newer version than the client, we send back an error and the client re-`exec`s itself using the server's binary -- except for `flow lsp`, which can't respawn itself since that would break its connection with the IDE.

if the server is an older version than the client, we kill the server and the client starts a new server, which in theory should be the newer version.

this works when *upgrading* flow, but breaks down if you *downgrade* flow -- for example, by checking out a commit using an older version while the server is running.

if the server was already running when downgrading:
- calling `flow status` sees that the server is newer so it respawns using the server's binary. so now you're running the higher-versioned binary on a commit that's supposed to use the downgraded version!
- calling `flow lsp` sees the server is newer, so it exits, hoping that the IDE will respawn `flow lsp` using the new binary -- but of course it respawns using the downgraded version, and errors again and again until the IDE gives up (5 crashes in 3 minutes in vscode).
- a long-running `flow lsp` will keep running. when the server notices a filesystem event like the .flowconfig changing, the server will exit. `flow lsp` will spawn the downgraded server, which will make `flow lsp` exit. vscode will respawn it with the correct binary. (this is good!)

instead, we should prefer whichever of the client or server was spawned more recently:

- calling `flow status` or `flow lsp` sees that it was spawned more recently than the mismatched `flow server`, so it kills the server and starts a new one using the downgraded binary
- a long-running `flow lsp` will behave like before: when the server exits, it'll start a downgraded one. it will see that the server was spawned more recently, so the server will survive and `flow lsp` will exit. when it is restarted by vscode, it'll be the downgraded binary and connect to the downgraded server.

Reviewed By: samwgoldman

Differential Revision: D35154665

fbshipit-source-id: e98523d127c35dbc99f7738616cd0c2f69e14aae