Re-register rolled-back-then-re-parsed files as file module providers
Summary:
If we parse a new file, we create a FileHeap entry F and a FileModuleHeap entry
M. We pass `new_file_provider = Some M` to `calc_dirty_modules`, which registers
F as a provider for M.

If we cancel this transaction before it commits, F's parse entity is rolled
back, but we leave the entry in FileHeap. We also remove F from M's providers.

Then, if we re-parse the same file, we will re-enter `add_checked_file` and hit
a slightly different code path. The FileHeap entry exists, but it's latest parse
is `None`.

In this case, we also need to pass `new_file_provider = Some M` to
`calc_dirty_modules`, to re-register F as a provider of M.

Reviewed By: mroch

Differential Revision: D35272981

fbshipit-source-id: cb92eef806c7ece3c04c0ffa904f6a5a239cd4eb