Remove unused "passthrough" flag for BindT
Summary:
I added this flag to support autobinding for React.createClass, where a spec
object might contain functions and non-functions, where functions should have a
bound this type, but everything else should be untouched.

Now createClass support is gone, we don't actually use this flag. All uses of
BindT do not have pass-through -- it is an error to try to bind things that
aren't functions.

Reviewed By: gkz

Differential Revision: D35274529

fbshipit-source-id: 7fa457532ae91f50a229bdb24f9bf9c4f2c90aa6