Drop exports from file sig
Summary:
The exports information used to be part of types-first, but is now subsumed by
type sigs.

Require information is used to build the dependency graph. Some more granular
information is used by the type normalizer and code actions, so I'm leaving this
alone for now.

We used to store the "source" information for `export ... from ...` statements
in the exports info, which I've moved to a new "require" kind, `ExportFrom`.

Reviewed By: mroch

Differential Revision: D35326928

fbshipit-source-id: d04a9c9cb98f910545984b714a99825d34754ba5