[new-env][EZ] Visit typeargs in new
Summary:
We are not visiting typeargs in new expressions at all. This caused them to be marked as unreachable and have empty type.

This problem seems to be carried over from SSA builder when we first forked it. This is fine for SSA builder since it doesn't care about types, but name_resolver does.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D35374667

fbshipit-source-id: e675bb6cba3d994331b42eb260aa975c45e598d3