handle partial reads in Marshal_tools.from_fd_with_preamble
Summary:
`Marshal_tools.from_fd_with_preamble` calls `Unix.read` to read the preamble. `read` is allowed to return less than the total number of bytes requested; you should call it again until it returns 0.

that's what `read_payload` does, so we use that for reading the preamble as well. this parallels how `write_payload` is used to write both the payload and the preamble.

this also means that if we fail to read the right number of bytes, then we hit end of file. my next diff will simplify the exceptions around this case.

Reviewed By: samwgoldman

Differential Revision: D35385988

fbshipit-source-id: 89e166b63988f4d70fb08a53ffdfb22a693c7bc9