[new-env] Precisely keep track of writes to global to distinguish writes to bound and unbound variables that share a name
Summary:
In D34948089 (https://github.com/facebook/flow/commit/d764a899d984e3eda5d62af04bd3601916634682), I introduced a fix that detects writes to globals in the new-env. The fix in that diff is not precise enough that it will confuse a legitimate write to a bound local variable to a write to global when they share the same name.

e.g.

```
unbound = 1; // Error

function actuallyBound() {
    let unbound;
    // No error, should not be confused with the real unbound name.
    unbound = 3;
}
```

This diff fixes this by tracking writes to global, and only do a builtin check if the location is a write to global instead of the name check.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D35365311

fbshipit-source-id: ce93918ff684e8b3c453904d5cf1a9e95456cd9e