Move resolved requires to new heap
Summary:
Instead of using the shared hash table for resolved requires lookups, this diff
makes resolved requires part of the file data structure in shared memory.

We only care about resolved requires for typed, parsed files, so the address of
resolved requires is stored along with typed parse information.

Resolved requires can change even if a file does not change. For example,
consider a file containing `require("foo")`. If a node module is created or
deleted, or a file with a different module extension, this might resolve to a
different file.

We accomodate this kind of change by making resolved requires an entity itself.
Note that parse information is also an entity, so we have nested entities.

While working on this diff, I noticed that the direct_dependent_files_cache has
a bug. In this test, we delete a file, then re-introduce the file and expect
no change in resolved requires. After this change, we see an "unexpected"
change.

It turns out, before this diff we failed to remove resolved requires when files
were deleted. Because of this, when the file was re-added, the old resolved
requires were still there, which appeared as unchanged.

Reviewed By: mroch

Differential Revision:
D35137880 (https://github.com/facebook/flow/commit/851f3372f30a29839c974d31081f068204ecd927)

------------------------------------------------------------------------
(from 095a75df61cbfc07ef10a0d4ee8fa77c59a7af71)

fbshipit-source-id: 99d02a8bd5d36a5098455b51aebae21df37c8b7e