Encode refinements propositionally and introduce changesets
Summary:
This diff changes the structure and implementation of refinements in the env builder, and introduces the notion of *changesets*, which are collections of new environment values whose lifetime is that of a refinement.

There are three major changes here, which are unfortunately hard to disentangle: we change the structure of the refinement maps that are applied to the environment to be keyed by ssa_id rather than lookup; we change the way that refinements are combined by representing them as propositions rather than as maps (but which are normalized to maps before being applied), and we build changesets for heap refinements.

## Refinement maps

Previously, the "live" refinements that had been applied to the environment were stored in a map from the `RefinementKey.lookup` representing the variable (something like `foo` or `foo.bar`) to a refinement ID (representing the refinement applied to the variable) and an SSA ID (indicating which actual version of the variable was refined).

With this diff, we change the structure of the map to be from the SSA id itself to the lookup name and the refinement id. In addition, the SSA id involved is no longer the immediate ID of the variable, but is instead the ID we find if we search through refinements. This change means that we no longer have to decide what to do if we see the same `lookup` refined in different ways with different IDs (as would happen if the variable was assigned to during a complex refinement), and we can be more consistent with what we do when we merge refinement scopes (previously, there was a violation of an implicit invariant when we merged refinement scopes, where we didn't consider if the SSA id was the same or not).

## Refinement propositions

Previously, refinement maps were essentially a large conjunction of various cases. If we had a case like `if (x && y) { } { } `, the result of inspecting the test would be a lookup map from `x` to `truthy` and `y` to `truthy`. Then, negating a refinement would just be adding `Not` in front of all the refinements in the map.

Of course, this turns out to be wrong in cases like the above: in the else case of this if, we shouldn't consider `x` to be falsey *and* `y` to be falsey, but rather it should be a disjunction. We could try to fix this by treating the negation of any map with more than one element in it to be the empty map, but that's not always right either! Consider:`if (x.y) { } { }`. In the then case, we have `x => PropExists(y)` and `x.y => Truthy`. In the else case, we should actually have `x => Not PropExists(y)` and `x.y => Falsey`! These two refinements should be understood as being linked biconditionally.

What we do instead is to actually encode refinements in a propositional form. Here, a base `Refinement` proposition is still a map from variables (in SSA id form, as above) to refinements, and when we negate this, we once again slap a `Not` in front of each of the refinements--everything in a single map is understood as being linked biconditionally. We can then connect multiple of these maps using conjunction and disjunction. So, `if (x.y)` creates the proposition `Refinement (x => PropExists(y), x.y => Truthy)`, and `if (x && y)` creates `(Refinement (x => Truthy)) /\ (Refinement (y => Truthy))`. Negating the former produces negative refinements on `x` and `x.y`, and negating the latter produces `(Not Refinement (x => Truthy)) \/ (Not Refinement (y => Truthy))`.

In order to actually apply these propositions to the environment, we still eventually have to come up with a single map representing specifically what concrete refinements are implied by the refinement proposition. We do so simply by recursively merging the refinement maps: when we see a conjunction, we take the union of the keys of the two operands, and with a disjunction, we take the intersection. Therefore the refinement proposition `(Not Refinement (x => Truthy)) \/ (Not Refinement (y => Truthy))` above is normalized to an empty map, since we don't gain definite information about either `x` or `y` from the refinement.

This is implemented in practice with a collection of helper API methods. To create a `Refinement` proposition of biconditionally linked refinements, we use `start_refinement` and `extend_refinement` to build up the overall map, before committing it (and conjuncting it with whatever refinement already exists in scope) by calling `add_refinement`. In cases where a single refinement is all that's needed, we can just call `add_single_refinement`, which works just like how `add_refinement` used to.

## Changesets

Finally, we observe that in some cases we don't just need to refine an environment entry within some refinement scope, we also need to create it, and perhaps remove it when we leave the scope. This arises in cases like

```
if (x && x.y) { } { }
```
In the then case, `x.y` should exist as a heap refinement and we should be able to access it, add other refinements to it, etc. However, in the else case, `x.y` may not actually exist! We shouldn't be able to look it up directly in the environment, but should instead have to do a traditional `GetProp` on `x` to see if it's there.

We accomplish this by means of "changesets", which are maps from lookup keys to Val.ts. Changesets are induced by refinements, though usually they don't have to be directly considered at the point when we're creating refinements. Instead, when we're calling `add_refinement` and applying a refinement to the environment using `map_val_with_lookup`, if we end up creating a new heap refinement with `create_val_for_heap`, we also save that val in a changeset. The changeset is considered to be an additional component of the `Refinement` proposition in the system described above, so they can be intersected and unioned by disjunction and conjunction similarly.

We also track the changeset that has currently been applied as part of the refinement scope, along with the overall proposition and the currently applied refinements. When we pop off a refinement scope, we remove the elements of the changeset from the environment, and when we push on a refinement scope, we reintroduce them to the environment as needed. So, in the above example, the refinement `x.y` produces a changeset containing `x.y`, the refinement `x` produces an empty changeset, and they're `And`ed together. In the then case, that gets normalized to a single changeset containing `x.y`, so in that scope we know that `x.y` exists in the environment. In the then case, we negate the proposition which turns the `And` into an `Or`, which results in a changeset with nothing in it when we intersect the keys--so x.y doesn't exist in the then case.

We have to treat logical op-assignment slightly specially here by explicitly creating a heap value for the LHS if it doesn't exist--previously this would be handled entirely by the vals added during refinement, but now those vals only live in the resulting refinement scope as part of the changeset. If we need a heap refinement to be "permanent", we need to add it explicitly.

Reviewed By: jbrown215

Differential Revision: D35322159

fbshipit-source-id: 7a9c390ebc805440fc02a7b234b1bea73e3e2253