Add codemod to add explicit type arguments to `useState` when it is called with an empty object
Summary:
Add codemod to add explicit type parameters to `useState` when it is called with an empty object. In the future this could be extended to also annotate calls with empty arrays, null, etc.

Changelog: [feature] Add codemod to add explicit type arguments to `useState` when it is called with an empty object

Reviewed By: panagosg7

Differential Revision: D34660654

fbshipit-source-id: b08e269b9a785970d24aad65d8f623353aea0063